template completed for article page
The green is a tab that has multiple posts or is auto scrolls through different articles. The red is going to be a more arrow that scrolls right through. Each of the three rows are different categories maybe. The blue box will be a 'scroll back to top' button. The very bottom is a footer!