fixed left side margin error
The div that styled all elements to the left 10% also included the footer. To fix this problem, I added the div left 10% pos absolute INSIDE each row element and not outside all elements. Although this takes up more code, it seems to be the most effective and readable way of going about this issue.